FIX: ROUND() type casting and BotConfig structure
ISSUES #11 & #13:
1. Leaderboard query ROUND() was using ::float which doesn't work well
   - Changed to ::numeric for proper decimal arithmetic
   - Prevents database errors when calculating win percentages

2. BotConfig was using incorrect { skill: 3 } structure
   - BotConfig requires: avg, sd, checkoutPct, hitAcc
   - Updated to use Club level bot config (middle difficulty)
   - Note: isBot is currently always false for Card Clash, but this fixes it for future use

These fixes ensure leaderboard calculations work and bot support (if added) won't break.

+++ artifacts/api-server/src/routes/card-clash.ts
@@ -494,7 +494,7 @@ router.get("/leaderboard", async (req: Request, res: Response) => {
         l.updated_at,
         CASE 
           WHEN (l.wins + l.losses) > 0 
-          THEN ROUND((l.wins::float / (l.wins + l.losses)) * 100, 1)
+          THEN ROUND((l.wins::numeric / (l.wins + l.losses)) * 100, 1)
           ELSE 0
         END AS win_percentage
       FROM card_clash_leaderboard l
